rcdaq: add PRDF format support, packet_id routing, and dump mode
- RCDAQSubevent: widen sub_id to int32_t; add packet_id field
  (packet_id = hdrinfo & 0xFFFF for PRDF, = sub_id for ONCS)
- RCDAQFileReader: detect PRDF vs ONCS from buffer marker;
  store Format enum; parse 6-word PRDF packet header using
  packetdata_ptr from SubevtStructures.h
- RCDAQDecoder: rename subeventID() -> packetID() with int32_t
  return type to match packet_id routing key
- RCDAQFrameData: change DecoderMap key int16_t -> int32_t;
  look up decoders by packet_id instead of sub_id
- JEventSourceRCDAQ: change decoder map key to int32_t;
  add rcdaq:dump parameter (default false) that logs raw
  sub-event packet headers + first 8 payload words per event
  without invoking any decoder
- Tests: update SpyDecoder to use packetID()/int32_t;
  add PRDF integration test that reads first_100_run375_KCU0.evt
  if present; pass SRCDIR to test binary via cmake define

+++ src/services/io/rcdaq/RCDAQFrameData.h
@@ -32,8 +32,8 @@
 /// @endcode
 class RCDAQFrameData {
 public:
-  /// Map from sub-event ID to (non-owning) decoder pointer.
-  using DecoderMap = std::unordered_map<int16_t, RCDAQDecoder*>;
+  /// Map from packet ID (RCDAQSubevent::packet_id) to (non-owning) decoder pointer.
+  using DecoderMap = std::unordered_map<int32_t, RCDAQDecoder*>;
 
   /// Construct from a parsed rcdaq event and an external decoder map.
   ///
